[refactor] 쿠키 유틸 분리~! #263
