+++ utils/axios.ts
@@ -1,21 +1,17 @@
 import axios from 'axios';
 
+import getAuthorization from 'utils/cookieUtil';
+
 const baseURL = `http://localhost:3000/api`;
-// const baseURL = 'http://localhost:10.19.223.86'; // 허남준
+// const baseURL = 'http://10.19.219.189:2998'; // 허남준
 
 const instance = axios.create({ baseURL });
 
 instance.interceptors.request.use(
   function setConfig(parameter) {
     const config = parameter;
-    const cookies: { [key: string]: string } = {};
 
-    document.cookie.split(';').forEach((cookie) => {
-      const [name, value] = cookie.trim().split('=');
-      cookies[name] = value;
-    });
-
-    const token = cookies.Authorization || null;
+    const token = getAuthorization();
 
     config.headers['Content-Type'] = 'application/json';
     if (token) config.headers.Authorization = `Bearer ${token}`;
